Arduino micros() is unreliable, replace with our own timer API.
This fixes a source of a problem that blocked the development for a
while.  Possibly due to the interrupts being disabled inside the
WS2811 output function or for some other reason micros() was
returning veyr unreliable results, lower than expected and I had
been looking for the cause of the calculations coming out wrong
pretty much everywhere except in the stock timer functions.  At about
300 iterations per second, micros() returned about half the real
value, and closer to 50 fps it was approaching 0.9 of the actual
value.  The values were a little random too, jumping up and down
about 5%.